// It actually works. Read for development fun-facts!!! >
As it turns out the token is an autonomous entity. If it was created with fix parameters, it doesn't matter if you overwrite the code and change,
let's say the type of the token. It won't matter if you check for the rigth type with the code.

>>> This in practice can mean, that you might think that the code is working, yet you are just testing with
an obsolete token.